Add files via upload
In this project we have mainly discussed about the pretraining of multilingual and Cross-lingual language model and also the difference between the monolingual and multilingual pretraining. it is possible to use cross-lingual models on various use cases, such as semantic search, 
plagiarism, and zero-shot text classification. we can train cross lingual model on a dataset from a langauge and test the model on a completely different language.
it is possible to use cross-lingual models on various use cases, such as semantic search, 
plagiarism, and zero-shot text classification